Map is fully rendered according to current world knowledge
All discovered rooms are drawn on the screen, with the current
room being centered in the window. This is done thanks to new
parameters in render() functions : offsets, that allow the map
to render rooms with a specified offset (in blocks).

Warning : there also exists an low-level offset (in PIXELS) enabling
all pixels drawn to be shifted (thus at the pixel level, in DrawImage
method) : VIEW_OFFSET_*. It is written for future uses.