Object uniqueness now treated when parsing file.
We ensure objects of the same room have unique identifiers when parsing the files describing maps rather than when the map is being created.

This will make it easier to save progress: we save the id of the object into the data files, and when the file is parsed again to load a saved file, we won't add integers to the end of the identifier repeatedly.