WIP : Save files, description of objects in a map can now be saved in a file.
Still need to take care of monsters, NPCs, and player inventory and health. Also, we need to add the player characteristics to the file parsing when creating a map.

Also we need to decide how the saveProgress function will be used in practise (after every warp to a new map and when a checkpoint is reached perhaps?).