WIP : Savefiles : Player characteristics can now be saved
Player characteristics can be saved, and working on inventory.
We will probably need to modify the structure of the inventory to facilitate accurate description of each object.
File parsing for map creation has also been modified : an exact description of how to initialise the player to the creator's convenience has been implemented and HowItWorks.txt has been updated accordingly.
Also, a default value is given to the player's initial position if it isn't specified.