Monster room translation
Fixed the problem of the monster removal in a for loop in Room
with Room::addMonsterToRemove.
Fixed coordinate translation.

Found a bug : when a monster follows a player in a room, and when
the player changes room and invokes another monster, the monster
left in the previous room appears next to the new monster immediately.

Several little graphical bugs remain, like quick disappear of a
monster right when it gets on the other side of a warp (due to
map design).