Map titles
To add to the folklore of the game, naming a map is now possible, by
adding said name just after the number of rooms in the format.
Note that this is a non breaking change in the format : older versions
of Doarm will still work with the new map format.

Fixed several bugs of colour blending with semi-transparent backgrounds
or foreground colours in text rendering.

Cleaned and limited code duplication in RenderContext.