Picked objects can be released
The user can right-click on a object in the inventory to
throw it away near its position in the current room.

For now the spot to throw objects in is constant, so a player
throwing away several objects will make a pile of objects.

Also the "worn" inventory now discriminate objects by
their address in the RAM (pointer value).