Player inventory saving and loading
The inventory is now saved when a checkpoint is encountered
by the player. It is initialised only at the beginning of the
game, through initialise and save functions, in Player.

The objects worn by the player are saved as references in the
inventory just above in the Start save file.

Player takes a new parameter when created : the name of the current
save slot.